--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomic/atoms/interfaces/AtomType.kt
@@ -1,8 +1,0 @@
-package me.kevinschildhorn.common.design.atomic.atoms.interfaces
-
-enum class AtomType {
-    BUTTON,
-    TEXT,
-    IMAGE,
-    TEXTFIELD
-}
