--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomic/atoms/interfaces/BorderedAtom.kt
@@ -1,7 +1,0 @@
-package me.kevinschildhorn.common.design.atomic.atoms.interfaces
-
-import me.kevinschildhorn.common.design.color.SharedColor
-
-interface BorderedAtom {
-    val borderColor: SharedColor
-}
