--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomic/molecules/BaseMolecule.kt
@@ -1,7 +1,0 @@
-package me.kevinschildhorn.common.design.atomic.molecules
-
-import me.kevinschildhorn.common.design.atomic.atoms.Atom
-
-abstract class BaseMolecule {
-    abstract val atoms: List<Atom>
-}
