+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/atoms/Atom.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.common.design.atomicLib.atoms
+
+import me.kevinschildhorn.common.design.atomicLib.atoms.interfaces.AtomType
+
+abstract class Atom {
+    abstract val type: AtomType
+}
