+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/atoms/interfaces/AtomType.kt
@@ -1,0 +1,8 @@
+package me.kevinschildhorn.common.design.atomicLib.atoms.interfaces
+
+enum class AtomType {
+    BUTTON,
+    TEXT,
+    IMAGE,
+    TEXTFIELD
+}
