+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/atoms/interfaces/BorderedAtom.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.common.design.atomicLib.atoms.interfaces
+
+import me.kevinschildhorn.common.design.atomicLib.color.SharedColor
+
+interface BorderedAtom {
+    val borderColor: SharedColor
+}
