+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/color/SharedColor.kt
@@ -1,0 +1,3 @@
+package me.kevinschildhorn.common.design.atomicLib.color
+
+expect class SharedColor(hex: Long)
