+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/font/SharedTypography.kt
@@ -1,0 +1,9 @@
+package me.kevinschildhorn.common.design.atomicLib.font
+
+enum class Weight {
+    NORMAL,
+    BOLD,
+    ITALIC
+}
+
+data class SharedTypography(val fontName: String, val weight: Weight, val size: Int)
