--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/color/SharedColor.kt
@@ -1,3 +1,0 @@
-package me.kevinschildhorn.common.design.color
-
-expect class SharedColor(hex: Long)
