--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/color/SharedEnabledColor.kt
@@ -1,6 +1,0 @@
-package me.kevinschildhorn.common.design.color
-
-data class SharedEnabledColor(
-    val color: SharedColor,
-    val disabledColor: SharedColor,
-)
