--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/font/SharedFont.kt
@@ -1,8 +1,0 @@
-package me.kevinschildhorn.common.design.font
-
-enum class Weight {
-    NORMAL,
-    BOLD,
-    ITALIC
-}
-class SharedFont(name: String, weight: Weight)
