--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/theme/DesignColors.kt
@@ -1,13 +1,0 @@
-package me.kevinschildhorn.common.design.theme
-
-import me.kevinschildhorn.common.design.color.SharedColor
-
-object DesignColors {
-    val primary = SharedColor(0xFF0000FF)
-    val secondary = SharedColor(0xFF808080)
-    val error = SharedColor(0xFFFF0000)
-    val success = SharedColor(0xFF00FF00)
-    val hint = SharedColor(0xFFBBBBBB)
-    val textColor = SharedColor(0xFF000000)
-    val invertedTextColor = SharedColor(0xFFFFFFFF)
-}
