--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/theme/DesignEnabledColors.kt
@@ -1,10 +1,0 @@
-package me.kevinschildhorn.common.design.theme
-
-import me.kevinschildhorn.common.design.color.SharedEnabledColor
-
-object DesignEnabledColors {
-    val default = SharedEnabledColor(
-        color = DesignColors.primary,
-        disabledColor = DesignColors.secondary
-    )
-}
