--- common/src/iosMain/kotlin/me/kevinschildhorn/common/design/theme/Fonts.kt
@@ -1,1 +1,0 @@
-package me.kevinschildhorn.common
